refactor(sequencer/config): Sequencer config moved to libs + correct usage in integration test

+++ libs/sdk/trigger-oracle/src/lib.rs
@@ -300,9 +300,9 @@ impl OracleTrigger {
     }
 }
 
-pub fn current_unix_time() -> u64 {
+pub fn current_unix_time() -> u128 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .expect("SystemTime before UNIX EPOCH!")
-        .as_secs()
+        .as_secs() as u128
 }
